CDAP-13098 add a starting program status
With the addition of the provisioning lifecycle, programs will be
in a pending state for minutes until their cluster is provisioned.
It is misleading to categorize this state as 'running', as users
may expect metrics to update and logs to be available, but they
really won't be ready for a few minutes. In order to better reflect
what is happening, added a starting program state instead of just
having running and stopped.

This helps with some unit tests that expect a program to be running
and not starting when the status endpoint returns running.

There is also a bug with stopping programs that are provisioning
but not yet running. This will be addressed in a future PR, but
until then this will fix the tests that fail if they happen to
stop a program while there is no program controller for it yet.