CDAP-13326 configure plugins at the start of a run
Call configurePipeline() on plugins at the start of a program
run right before prepareRun() is called. This is done so that
anything that could not be done when the pipeline was deployed
due to unevaluated macros will be done at the start of the
pipeline run. This means any datasets with macro evaluated
names will be created, macro dependent schema will be
propagated, and macro dependent validations will be executed.

Moved the schema propagation logic from the PipelineSpecGenerator
into its own class that can be used at runtime by ETLMapReduce
and ETLSpark.