CDAP-13326 configure plugins at runtime
Configuring plugins at runtime allows the app to propagate
any schema that could not be determined at deploy time but which
could be determined at runtime after macro evaluation.

This removes a lot of duplicate code in plugins around schema
fetching, validation, and dataset creation.

Before the prepareRun() method for a plugin is called, its
configurePipeline() method will be called. The updated phase
spec will be written out to the Hadoop Configuration or Spark
config so that the updated schemas will be available to
mappers/reducers/executors.

The main challenge is propagating schema across pipeline phases.
This is done by writing the relevant schema information in the
workflow token when it needs to traverse phase boundaries.