CDAP-16852 handle dynamic schemas in auto join
Properly handle the case when the inputs into an auto-join stage
have null schemas due to the schema being unknown at deploy time,
usually because of macros.

In order to support this type of use case, added a way for a
plugin to specify the output schema in the JoinDefinition.
This should be used when the output schema cannot be derived
at deployment time due to macros. This is what the existing
Joiner currently requires the user to do for dynamic join
use cases.

For Spark, the implementation becomes significantly more difficult
due to the fact that the schema of each input stage needs to be
known in the Spark driver, in order to convert RDDs into
DataFrames.

Added logic that derives the input schema using the output schema,
selected fields, and join keys. It is possible to derive
a usable schema when all the join keys are present in the final
output schema.