CDAP-16875 include more detailed information in join exception
Added a list of JoinErrors to the InvalidJoinException and
refactored the join validation logic to include all the
errors found in that list. This allows plugins to
translate the errors found by the platform to problems
with plugin specific configuration properties.
Without this, plugins would have to re-implement all of the
validation checks that are performed by the platform.

Enhance the AutoJoinerContext to include a way to get a
FailureCollector to allow plugins to handle validation of the
join definition in the same way as other validation errors.