CDAP-16935 honor partitions set by autojoin plugins
Unfortunately the contract cannot be honored fully due to the
fact that Spark dataframe join partitions are controlled by
a setting in the spark conf, which means they must be constant
for everything within the same Spark program.

If multiple joiners specify different values, the highest value
is chosen and a warning log message is emitted for each joiner
whose partition value was ignored.