CDAP-17097 add back port schemas to stage spec to fix validate issue
StageSpec used to contain a map of ports to their schemas. This
map is never used by the backend, but is returned as part of the
stage validation call. This is in turn used by the UI to generate
the set of output ports along with their schemas.

Added back the map to return the API to compatibility and fix the
UI.