Fixes initial load of UI to serve correctly in k8s environment
  - Adds handling SIGTERM signal to gracefull exit nodejs app
  - Fixes initial loading of CDAP UI in k8s environment after a pod restart

The root cause this happened only the first time is,

- When we start the UI server we parse the cdap config and store it in memory
- The node server now starts to listen for any ws connections from client
- Once UI receives a ws connection we instantiate the Aggregator module to make rest calls
- This again requires cdap config which we again generate.
- We start to listen to ws messages only after the cdap config is extracted
- Between parsing the config and starting to listen, if node server receives any messages
  from any ws client it just drops them.

This is why the first few requests don't reach the ws server started by nodejs server.

The fix is to avoid generating cdap config multiple times and start to listen to ws
messages as soon as a client ws connection is established.